+++ recap/readers/hive/thrift_gen/ThriftHiveMetastore.py
