+++ recap/readers/hive/thrift_gen/constants.py
