+++ recap/readers/hive/thrift_gen/ttypes.py
